# [MacAddress]: Optimize to_string() and parseMacString methods