MacAddress class instantiated many times from string. Also we parse mac addresses a lot. I did a benchmark and I found that our old version of to_string() and parseMacString() methods not so fast.

I've rewritten both of them. Now both methods faster (parseMacString - 76 times faster, almost 10 times faster) and parseMacString() catch more errors.

```
2018-01-03 23:01:53
Run on (16 X 2394.44 MHz CPU s)
CPU Caches:
  L1 Data 32K (x16)
  L1 Instruction 32K (x16)
  L2 Unified 256K (x16)
  L3 Unified 30720K (x16)
------------------------------------------------------------
Benchmark                     Time           CPU Iterations
------------------------------------------------------------
BM_ParseMacOld              840 ns        840 ns     842880
BM_ParseMac                  11 ns         11 ns   62703966
BM_MAC_to_string_old        496 ns        496 ns    1448299
BM_MAC_to_string             50 ns         50 ns   13564922
```